requirements.txt
python-telegram-bot==4.0rc1
pyvndb
